LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/moveBefore-option-recalc-style-expected.txt



FAIL moveBefore() of an option element between two selects assert_equals: expected Element node <select id="move_to"></select> but got Element node <select>
  <option id="mv">OPTION</option>
</select>
